chore: change files named types to avoid potential issues
