+++ stoix/networks/base.py
@@ -8,9 +8,9 @@ import jax.numpy as jnp
 import numpy as np
 from flax import linen as nn
 
+from stoix.base_types import Observation, RNNObservation
 from stoix.networks.inputs import ObservationInput
 from stoix.networks.utils import parse_rnn_cell
-from stoix.types import Observation, RNNObservation
 
 
 class FeedForwardActor(nn.Module):
