+++ stoix/systems/ppo/ff_ppo.py
@@ -14,6 +14,7 @@ from jumanji.env import Environment
 from omegaconf import DictConfig, OmegaConf
 from rich.pretty import pprint
 
+from stoix.base_types import ActorApply, CriticApply, ExperimentOutput, LearnerFn
 from stoix.evaluator import evaluator_setup, get_distribution_act_fn
 from stoix.networks.base import FeedForwardActor as Actor
 from stoix.networks.base import FeedForwardCritic as Critic
@@ -23,7 +24,6 @@ from stoix.systems.ppo.types import (
     LearnerState,
     PPOTransition,
 )
-from stoix.types import ActorApply, CriticApply, ExperimentOutput, LearnerFn
 from stoix.utils import make_env as environments
 from stoix.utils.checkpointing import Checkpointer
 from stoix.utils.jax_utils import (
