+++ stoix/systems/sac/ff_sac.py
@@ -17,18 +17,18 @@ from jumanji.types import TimeStep
 from omegaconf import DictConfig, OmegaConf
 from rich.pretty import pprint
 
+from stoix.base_types import ActorApply, ExperimentOutput, LearnerFn, LogEnvState
 from stoix.evaluator import evaluator_setup, get_distribution_act_fn
 from stoix.networks.base import CompositeNetwork
 from stoix.networks.base import FeedForwardActor as Actor
 from stoix.networks.base import MultiNetwork
 from stoix.systems.q_learning.types import QsAndTarget, Transition
-from stoix.systems.sac.types import (
+from stoix.systems.sac.sac_types import (
     ContinuousQApply,
     SACLearnerState,
     SACOptStates,
     SACParams,
 )
-from stoix.types import ActorApply, ExperimentOutput, LearnerFn, LogEnvState
 from stoix.utils import make_env as environments
 from stoix.utils.checkpointing import Checkpointer
 from stoix.utils.jax_utils import unreplicate_batch_dim, unreplicate_n_dims
