+++ stoix/systems/search/ff_az.py
@@ -17,6 +17,7 @@ from jumanji.env import Environment
 from omegaconf import DictConfig, OmegaConf
 from rich.pretty import pprint
 
+from stoix.base_types import ActorApply, CriticApply, ExperimentOutput, LearnerFn
 from stoix.networks.base import FeedForwardActor as Actor
 from stoix.networks.base import FeedForwardCritic as Critic
 from stoix.systems.ppo.types import (
@@ -25,13 +26,12 @@ from stoix.systems.ppo.types import (
     LearnerState,
 )
 from stoix.systems.search.evaluator import search_evaluator_setup
-from stoix.systems.search.types import (
+from stoix.systems.search.search_types import (
     EnvironmentStep,
     ExItTransition,
     RootFnApply,
     SearchApply,
 )
-from stoix.types import ActorApply, CriticApply, ExperimentOutput, LearnerFn
 from stoix.utils import make_env as environments
 from stoix.utils.checkpointing import Checkpointer
 from stoix.utils.jax_utils import (
