+++ stoix/systems/vpg/ff_reinforce_continuous.py
@@ -15,6 +15,7 @@ from jumanji.env import Environment
 from omegaconf import DictConfig, OmegaConf
 from rich.pretty import pprint
 
+from stoix.base_types import ActorApply, CriticApply, ExperimentOutput, LearnerFn
 from stoix.evaluator import evaluator_setup, get_distribution_act_fn
 from stoix.networks.base import FeedForwardActor as Actor
 from stoix.networks.base import FeedForwardCritic as Critic
@@ -23,8 +24,7 @@ from stoix.systems.ppo.types import (
     ActorCriticParams,
     LearnerState,
 )
-from stoix.systems.vpg.types import Transition
-from stoix.types import ActorApply, CriticApply, ExperimentOutput, LearnerFn
+from stoix.systems.vpg.vpg_types import Transition
 from stoix.utils import make_env as environments
 from stoix.utils.checkpointing import Checkpointer
 from stoix.utils.jax_utils import unreplicate_batch_dim, unreplicate_n_dims
