+++ stoix/wrappers/gymnax.py
@@ -12,7 +12,7 @@ from jumanji.specs import Array, DiscreteArray, Spec
 from jumanji.types import StepType, TimeStep, restart
 from jumanji.wrappers import Wrapper
 
-from stoix.types import Observation
+from stoix.base_types import Observation
 
 if TYPE_CHECKING:  # https://github.com/python/mypy/issues/6239
     from dataclasses import dataclass
